requirements.txt
wavio
speech_recognition
sounddevice
scipy.io.wavfile
pynput
hugchat
playsound
gTTS
